Rename HypothesisChromeExtension => Extension
Rename HypothesisChromeExtension to Extension and convert it to a class.

 - Simplify the name

 - Convert from a default to named export, per our current coding
   conventions

 - Convert from ES5-style class to a modern class
